deleteing tests in processor and modifying shifts in plugin
