+++ modernizer-maven-plugin/src/test/java/org/gaul/modernizer_maven_plugin/ModernizerTest.java
@@ -18,12 +18,13 @@ package org.gaul.modernizer_maven_plugin;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.io.BufferedReader;
 import java.io.File;
+import java.io.FileReader;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
-import java.io.UnsupportedEncodingException;
 import java.net.HttpURLConnection;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
@@ -31,11 +32,9 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
-import java.util.HashSet;
 import java.util.Hashtable;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.SortedMap;
 import java.util.Vector;
 import java.util.regex.Pattern;
@@ -94,6 +93,9 @@ public final class ModernizerTest {
             Collections.<String>emptySet();
     private static final Collection<String> NO_IGNORED_METHODS =
             Collections.<String>emptySet();
+    private List<Pattern> ignoreClasses = new ArrayList<Pattern>();
+    private List<String> ignoreMethods = new ArrayList<String>();
+
 
     @Before
     public void setUp() throws Exception {
@@ -103,6 +105,24 @@ public final class ModernizerTest {
             violations = Modernizer.parseFromXml(is);
         } finally {
             Utils.closeQuietly(is);
+        }
+        File currentDirectory = new File(System.getProperty("user.dir"));
+        File ignoreClassesFile = new File(currentDirectory,
+            "/target/modernizer/test/ignore-annotated-classes.txt");
+        File ignoreMethodsFile = new File(currentDirectory,
+            "/target/modernizer/test/ignore-annotated-methods.txt");
+
+        BufferedReader br =
+            new BufferedReader(new FileReader(ignoreClassesFile));
+        String line;
+        while ((line = br.readLine()) != null) {
+            ignoreClasses.add(Pattern.compile(line));
+        }
+
+        br = new BufferedReader(new FileReader(ignoreMethodsFile));
+        ignoreMethods = new ArrayList<String>();
+        while ((line = br.readLine()) != null) {
+            ignoreMethods.add(line);
         }
     }
 
@@ -387,32 +407,22 @@ public final class ModernizerTest {
 
     @Test
     public void checkIgnoreMethodWithEmptyParameters() throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreMethodWithEmptyParametersTestClass.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-                "$IgnoreMethodWithEmptyParametersTestClass " +
-                "testMethodEmptyParameters void");
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreMethodWithEmptyParametersTestClass.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(0);
     }
 
     @Test
     public void checkIgnoreMethodWithVoidParameter() throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreMethodWithVoidParameterTestClass.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-                "$IgnoreMethodWithVoidParameterTestClass " +
-                "testMethodVoidParameter void java.lang.Void");
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreMethodWithVoidParameterTestClass.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(0);
     }
@@ -420,67 +430,44 @@ public final class ModernizerTest {
     @Test
     public void checkIgnoreMethodWithPrimitiveTypeParameters()
         throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreMethodWithPrimitiveTypeParametersTestClass.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-            "$IgnoreMethodWithPrimitiveTypeParametersTestClass " +
-            "testMethodPrimitiveTypeParameters void int boolean byte char " +
-            "short long float double");
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreMethodWithPrimitiveTypeParametersTestClass.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(0);
     }
 
     @Test
     public void checkIgnoreOverloadedConstructor() throws Exception {
-        ClassReader cr =
-            new ClassReader(IgnoreConstructorTestClass.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-            "$IgnoreConstructorTestClass <init> void");
+        ClassReader cr = new ClassReader(
+            ModernizerTestHelper.IgnoreConstructorTestClass.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(1);
     }
 
     @Test
     public void checkIgnoreMethodWithGenericTypeParameters() throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreMethodWithGenericTypeParametersTestClass.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-            "$IgnoreMethodWithGenericTypeParametersTestClass " +
-            "testMethodMultipleGenericTypeParameters " +
-            "void java.util.List java.util.Set");
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreMethodWithGenericTypeParametersTestClass.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(0);
     }
 
     @Test
     public void checkIgnoreMethodWithDeclaredTypeParameter() throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreMethodWithDeclaredTypeParametersTestClass.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-            "$IgnoreMethodWithDeclaredTypeParametersTestClass " +
-            "testMethodDeclaredTypeParameter " +
-            "void org.gaul.modernizer_maven_plugin.ModernizerTest" +
-            ".StringGetBytesString");
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreMethodWithDeclaredTypeParametersTestClass.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(0);
     }
@@ -488,116 +475,78 @@ public final class ModernizerTest {
     @Test
     public void checkIgnoreMethodWithPrimitiveAndGenericParameters()
         throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreMethodWithPrimitiveAndGenericTypeParametersTestClass.class
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreMethodWithPrimitiveAndGenericTypeParametersTestClass.class
                 .getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-            "$IgnoreMethodWithPrimitiveAndGenericTypeParametersTestClass " +
-            "testMethodPrimitiveAndGenericTypeParameters " +
-            "void java.util.List boolean");
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(0);
     }
 
     @Test
     public void checkIgnoreOverloadedMethod() throws Exception {
-        ClassReader cr =
-            new ClassReader(IgnoreOverloadedMethodTestClass.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-            "$IgnoreOverloadedMethodTestClass testOverloadedMethod void");
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreOverloadedMethodTestClass.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(1);
     }
 
     @Test
     public void checkIgnoreMethodWithArrayTypeParameters() throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreMethodWithArrayTypeParametersTestClass.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-            "$IgnoreMethodWithArrayTypeParametersTestClass " +
-            "testArrayParameters " +
-            "void int[] java.lang.String[] float[][] java.util.List[] " +
-            "org.gaul.modernizer_maven_plugin.ModernizerTest" +
-            ".IgnoreOverloadedMethodTestClass[]");
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreMethodWithArrayTypeParametersTestClass.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(0);
     }
 
     @Test
     public void checkIgnoreGenericClassConstructor() throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreGenericClassConstructorTestClass.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-            "$IgnoreGenericClassConstructorTestClass " +
-            "<init> void");
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreGenericClassConstructorTestClass.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(1);
     }
 
     @Test
     public void checkIgnoreMethodInGenericClass() throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreMethodInGenericClassTest.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-            "$IgnoreMethodInGenericClassTest " +
-            "testGenericMethod void java.lang.Object");
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreMethodInGenericClassTest.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(0);
     }
 
     @Test
     public void checkIgnoreOverloadedMethodInGenericClass() throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreOverloadedMethodInGenericClassTest.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-            "$IgnoreOverloadedMethodInGenericClassTest " +
-            "testGenericOverloadedMethod void int java.lang.Object " +
-            "java.util.List java.util.List[]");
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreOverloadedMethodInGenericClassTest.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(1);
     }
 
     @Test
     public void checkIgnoreGenericClass() throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreGenericClass.class.getName());
-        Set<Pattern> ignoreClassName = new HashSet<Pattern>();
-        ignoreClassName.add(Pattern.compile(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-            "\\$IgnoreGenericClass"));
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreGenericClass.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, ignoreClassName, NO_IGNORED_METHODS)
+            NO_IGNORED_PACKAGES, ignoreClasses, NO_IGNORED_METHODS)
             .check(cr);
         assertThat(occurences).hasSize(0);
     }
@@ -605,18 +554,11 @@ public final class ModernizerTest {
     @Test
     public void checkIgnoreMethodReturningArrayOfDeclaredType()
         throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreMethodReturningArrayClassTest.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-                "$IgnoreMethodReturningArrayClassTest " +
-                "testMethodReturningArray " +
-                "org.gaul.modernizer_maven_plugin.ModernizerTest" +
-                ".IgnoreOverloadedMethodTestClass[]");
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreMethodReturningArrayClassTest.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(0);
     }
@@ -624,69 +566,44 @@ public final class ModernizerTest {
     @Test
     public void checkIgnoreMethodReturningArrayOfPrimitiveType()
         throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreMethodReturningArrayPrimitiveTypeClassTest.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-                "$IgnoreMethodReturningArrayPrimitiveTypeClassTest " +
-                "testMethodReturningArrayPrimitveType " +
-                "int[]");
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreMethodReturningArrayPrimitiveTypeClassTest.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(0);
     }
 
     @Test
     public void checkIgnoreMethodReturningPrimitiveType() throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreMethodReturningPrimitiveTypeClassTest.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-                "$IgnoreMethodReturningPrimitiveTypeClassTest " +
-                "testMethodReturningPrimitiveType " +
-                "char");
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreMethodReturningPrimitiveTypeClassTest.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(0);
     }
 
     @Test
     public void checkIgnoreMethodReturningDeclaredType() throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreMethodReturningDeclaredTypeClassTest.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-                "$IgnoreMethodReturningDeclaredTypeClassTest " +
-                "testMethodReturningDeclaredType " +
-                "org.gaul.modernizer_maven_plugin" +
-                ".ModernizerTest.IgnoreConstructorTestClass");
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreMethodReturningDeclaredTypeClassTest.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(0);
     }
 
     @Test
     public void checkIgnoreMethodReturningGenericType() throws Exception {
-        ClassReader cr = new ClassReader(
-            IgnoreMethodReturningGenericTypeClassTest.class.getName());
-        Set<String> ignoreMethodName = new HashSet<String>();
-        ignoreMethodName.add(
-            "org/gaul/modernizer_maven_plugin/ModernizerTest" +
-                "$IgnoreMethodReturningGenericTypeClassTest " +
-                "testMethodReturningGenericType " +
-                "java.util.List");
+        ClassReader cr = new ClassReader(ModernizerTestHelper
+            .IgnoreMethodReturningGenericTypeClassTest.class.getName());
         Collection<ViolationOccurrence> occurences = new Modernizer(
             "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethodName)
+            NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS, ignoreMethods)
             .check(cr);
         assertThat(occurences).hasSize(0);
     }
@@ -951,179 +868,5 @@ public final class ModernizerTest {
     }
 
     private enum EnumClass {
-    }
-
-    private static class IgnoreMethodWithEmptyParametersTestClass {
-        private static void testMethodEmptyParameters() throws Exception {
-            "".getBytes("UTF-8");
-        }
-    }
-
-    private static class IgnoreMethodWithVoidParameterTestClass {
-        private static void testMethodVoidParameter(Void var) throws Exception {
-            "".getBytes("UTF-8");
-        }
-    }
-
-    private static class IgnoreMethodWithPrimitiveTypeParametersTestClass {
-        private static void testMethodPrimitiveTypeParameters(
-            int intVar,
-            boolean boolVar,
-            byte byteVar,
-            char charVar,
-            short shortVar,
-            long longVar,
-            float floatVar,
-            double doubleVar
-        ) throws Exception {
-            "".getBytes("UTF-8");
-        }
-    }
-
-    private static class IgnoreConstructorTestClass {
-        IgnoreConstructorTestClass() throws Exception {
-            "".getBytes("UTF-8");
-        }
-
-        IgnoreConstructorTestClass(String string) throws Exception {
-            "".getBytes("UTF-8");
-        }
-    }
-
-    private static class IgnoreMethodWithGenericTypeParametersTestClass {
-        private static void testMethodMultipleGenericTypeParameters(
-            List<String> stringList,
-            Set<Integer> integerSet
-        ) throws Exception {
-            "".getBytes("UTF-8");
-        }
-    }
-
-    private static class IgnoreMethodWithDeclaredTypeParametersTestClass {
-        private static void testMethodDeclaredTypeParameter(
-            StringGetBytesString obj
-        ) throws Exception {
-            "".getBytes("UTF-8");
-        }
-    }
-
-    private static class
-        IgnoreMethodWithPrimitiveAndGenericTypeParametersTestClass {
-        private static void testMethodPrimitiveAndGenericTypeParameters(
-            List<String> list,
-            boolean bool
-        ) throws Exception {
-            "".getBytes("UTF-8");
-        }
-    }
-
-    private static class IgnoreOverloadedMethodTestClass {
-        private static void testOverloadedMethod() throws Exception {
-            "".getBytes("UTF-8");
-        }
-
-        private static void testOverloadedMethod(
-            String string
-        ) throws Exception {
-            string.getBytes("UTF-8");
-        }
-    }
-
-    private static class IgnoreMethodWithArrayTypeParametersTestClass {
-        private static void testArrayParameters(
-            int[] array,
-            String[] stringArray,
-            float[][] arrayOfArrays,
-            List<String> [] listOfArrayOfStrings,
-            IgnoreOverloadedMethodTestClass[] obj
-        ) throws Exception {
-            "".getBytes("UTF-8");
-        }
-    }
-
-    private static class IgnoreGenericClassConstructorTestClass<E> {
-        IgnoreGenericClassConstructorTestClass() throws Exception {
-            "".getBytes("UTF-8");
-        }
-        IgnoreGenericClassConstructorTestClass(
-            IgnoreOverloadedMethodTestClass ignore
-        ) throws Exception {
-            "".getBytes("UTF-8");
-        }
-    }
-
-    private static class IgnoreMethodInGenericClassTest<E> {
-        public void testGenericMethod(E obj) throws Exception {
-            "".getBytes("UTF-8");
-        }
-
-    }
-
-    public static class IgnoreOverloadedMethodInGenericClassTest<E> {
-        public final void testGenericOverloadedMethod(
-            E firstObj,
-            E secondObj
-        ) throws Exception {
-            "".getBytes("UTF-8");
-        }
-
-        public final void testGenericOverloadedMethod(
-            int var,
-            E genericVar,
-            List<E> list,
-            List<E>[] lists
-        ) throws Exception {
-            "".getBytes("UTF-8");
-        }
-    }
-
-    public static class IgnoreGenericClass<E> {
-        public final void testMethodInGenericClass() throws Exception {
-            "".getBytes("UTF-8");
-        }
-    }
-
-    public static class IgnoreMethodReturningArrayClassTest {
-        public final IgnoreOverloadedMethodTestClass[]
-            testMethodReturningArray() throws UnsupportedEncodingException {
-            IgnoreOverloadedMethodTestClass[] ex =
-                new IgnoreOverloadedMethodTestClass[1];
-            "".getBytes("UTF-8");
-            return ex;
-        }
-    }
-
-    public static class IgnoreMethodReturningArrayPrimitiveTypeClassTest {
-        public final int[] testMethodReturningArrayPrimitveType()
-            throws UnsupportedEncodingException {
-            int[] intArray = new int[1];
-            "".getBytes("UTF-8");
-            return intArray;
-        }
-    }
-
-    public static class IgnoreMethodReturningPrimitiveTypeClassTest {
-        public final char testMethodReturningPrimitiveType()
-            throws UnsupportedEncodingException {
-            "".getBytes("UTF-8");
-            return ' ';
-        }
-    }
-
-    public static class IgnoreMethodReturningDeclaredTypeClassTest {
-        public final IgnoreConstructorTestClass
-            testMethodReturningDeclaredType()
-            throws Exception {
-            "".getBytes("UTF-8");
-            return new IgnoreConstructorTestClass();
-        }
-    }
-
-    public static class IgnoreMethodReturningGenericTypeClassTest {
-        public final List<List<String>> testMethodReturningGenericType()
-            throws Exception {
-            "".getBytes("UTF-8");
-            return new ArrayList<List<String>>();
-        }
     }
 }
